xong tat ca

+++ QLNH/src/Display/TrangNhanVien.java
@@ -12,6 +12,7 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
+import model.NhanVien;
 
 /**
  *
@@ -25,11 +26,12 @@ public class TrangNhanVien extends javax.swing.JFrame {
     DefaultTableModel model;
     Connection cn;
     int index;
-
+    ArrayList<NhanVien> lstnv = new ArrayList<>();
     public TrangNhanVien() {
         initComponents();
         setLocationRelativeTo(null);
         cn = Helper.ketnoi("QLNH");
+       setDefaultCloseOperation(DISPOSE_ON_CLOSE);
 //        if(cn!= null){
 //            JOptionPane.showMessageDialog(this, "Dang nhap thanh cong");
 //        }else{
@@ -37,6 +39,8 @@ public class TrangNhanVien extends javax.swing.JFrame {
 //        }
         model = new DefaultTableModel();
         model = (DefaultTableModel) tblnv.getModel();
+        loaddata();
+       
         loadtable();
 
     }
@@ -53,15 +57,15 @@ public class TrangNhanVien extends javax.swing.JFrame {
         jPanel4 = new javax.swing.JPanel();
         jLabel1 = new javax.swing.JLabel();
         jPanel3 = new javax.swing.JPanel();
-        jButton22 = new javax.swing.JButton();
-        jButton23 = new javax.swing.JButton();
-        jButton24 = new javax.swing.JButton();
-        jButton25 = new javax.swing.JButton();
+        btnupdate = new javax.swing.JButton();
+        btnadd = new javax.swing.JButton();
+        btnxoa = new javax.swing.JButton();
+        btndoimk = new javax.swing.JButton();
         txttim = new javax.swing.JTextField();
         jButton6 = new javax.swing.JButton();
         jScrollPane1 = new javax.swing.JScrollPane();
         tblnv = new javax.swing.JTable();
-        jLabel3 = new javax.swing.JLabel();
+        lblcurrrent = new javax.swing.JTextField();
         jLabel2 = new javax.swing.JLabel();
         jPanel1 = new javax.swing.JPanel();
         jButton1 = new javax.swing.JButton();
@@ -84,39 +88,39 @@ public class TrangNhanVien extends javax.swing.JFrame {
 
         jPanel3.setOpaque(false);
 
-        jButton22.setBackground(new java.awt.Color(35, 35, 35));
-        jButton22.setForeground(new java.awt.Color(255, 255, 255));
-        jButton22.setText("Chỉnh Sửa");
-        jButton22.addActionListener(new java.awt.event.ActionListener() {
+        btnupdate.setBackground(new java.awt.Color(35, 35, 35));
+        btnupdate.setForeground(new java.awt.Color(255, 255, 255));
+        btnupdate.setText("Chỉnh Sửa");
+        btnupdate.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton22ActionPerformed(evt);
+                btnupdateActionPerformed(evt);
             }
         });
 
-        jButton23.setBackground(new java.awt.Color(35, 35, 35));
-        jButton23.setForeground(new java.awt.Color(255, 255, 255));
-        jButton23.setText("Thêm Mới");
-        jButton23.addActionListener(new java.awt.event.ActionListener() {
+        btnadd.setBackground(new java.awt.Color(35, 35, 35));
+        btnadd.setForeground(new java.awt.Color(255, 255, 255));
+        btnadd.setText("Thêm Mới");
+        btnadd.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton23ActionPerformed(evt);
+                btnaddActionPerformed(evt);
             }
         });
 
-        jButton24.setBackground(new java.awt.Color(35, 35, 35));
-        jButton24.setForeground(new java.awt.Color(255, 255, 255));
-        jButton24.setText("Xóa");
-        jButton24.addActionListener(new java.awt.event.ActionListener() {
+        btnxoa.setBackground(new java.awt.Color(35, 35, 35));
+        btnxoa.setForeground(new java.awt.Color(255, 255, 255));
+        btnxoa.setText("Xóa");
+        btnxoa.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton24ActionPerformed(evt);
+                btnxoaActionPerformed(evt);
             }
         });
 
-        jButton25.setBackground(new java.awt.Color(35, 35, 35));
-        jButton25.setForeground(new java.awt.Color(255, 255, 255));
-        jButton25.setText("Đổi Mật Khẩu");
-        jButton25.addActionListener(new java.awt.event.ActionListener() {
+        btndoimk.setBackground(new java.awt.Color(35, 35, 35));
+        btndoimk.setForeground(new java.awt.Color(255, 255, 255));
+        btndoimk.setText("Đổi Mật Khẩu");
+        btndoimk.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton25ActionPerformed(evt);
+                btndoimkActionPerformed(evt);
             }
         });
 
@@ -126,23 +130,23 @@ public class TrangNhanVien extends javax.swing.JFrame {
             jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel3Layout.createSequentialGroup()
                 .addContainerGap()
-                .addComponent(jButton23, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(btnadd, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(41, 41, 41)
-                .addComponent(jButton24, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(btnxoa, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(42, 42, 42)
-                .addComponent(jButton22, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(btnupdate, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(45, 45, 45)
-                .addComponent(jButton25, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(btndoimk, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap(180, Short.MAX_VALUE))
         );
         jPanel3Layout.setVerticalGroup(
             jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel3Layout.createSequentialGroup()
                 .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jButton23, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jButton24, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jButton22, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jButton25, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(btnadd, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnxoa, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnupdate, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btndoimk, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addGap(0, 2, Short.MAX_VALUE))
         );
 
@@ -165,14 +169,19 @@ public class TrangNhanVien extends javax.swing.JFrame {
                 "STT", "Mã nhân viên", "Họ tên", "Ngày sinh", "Điện thoại", "Chức vụ", "Số cmt"
             }
         ));
+        tblnv.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                tblnvMouseClicked(evt);
+            }
+        });
         jScrollPane1.setViewportView(tblnv);
 
         jPanel4.add(jScrollPane1, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 120, 800, 340));
 
-        jLabel3.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        jLabel3.setForeground(new java.awt.Color(255, 255, 255));
-        jLabel3.setText("Tổng số bản ghi: ");
-        jPanel4.add(jLabel3, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 470, 140, 30));
+        lblcurrrent.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
+        lblcurrrent.setForeground(new java.awt.Color(255, 0, 0));
+        lblcurrrent.setText("Tổng số bản ghi:");
+        jPanel4.add(lblcurrrent, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 473, 170, 30));
 
         jLabel2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/image/nen.jpg"))); // NOI18N
         jLabel2.setPreferredSize(new java.awt.Dimension(836, 587));
@@ -311,7 +320,7 @@ public class TrangNhanVien extends javax.swing.JFrame {
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
-    private void jButton22ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton22ActionPerformed
+    private void btnupdateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnupdateActionPerformed
         // TODO add your handling code here:
         try {
             int index = tblnv.getSelectedRow();
@@ -332,22 +341,29 @@ public class TrangNhanVien extends javax.swing.JFrame {
         }
 
 
-    }//GEN-LAST:event_jButton22ActionPerformed
+    }//GEN-LAST:event_btnupdateActionPerformed
 
-    private void jButton23ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton23ActionPerformed
+    private void btnaddActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnaddActionPerformed
         // TODO add your handling code here:
         new ThemNV().setVisible(true);
         model.setRowCount(0);
         loadtable();
 
-    }//GEN-LAST:event_jButton23ActionPerformed
+    }//GEN-LAST:event_btnaddActionPerformed
 
-    private void jButton24ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton24ActionPerformed
+    private void btnxoaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnxoaActionPerformed
         // TODO add your handling code here:
 
         try {
             int index = tblnv.getSelectedRow();
-            //xoa trong csdl va bang 
+            //xoa trong csdl va bang
+            if(index >0){
+                 int hoi = JOptionPane.showConfirmDialog(this, "Bạn muốn xóa mã nhân viên " + tblnv.getValueAt(index, 1));
+            if(hoi != JOptionPane.YES_NO_OPTION){
+                return;
+            }
+            }
+           
             String sql = "delete from nhanvien\n"
                     + "where manv=?";
 
@@ -362,13 +378,13 @@ public class TrangNhanVien extends javax.swing.JFrame {
             JOptionPane.showMessageDialog(this, "Khong xoa dc");
             e.printStackTrace();
         }
-    }//GEN-LAST:event_jButton24ActionPerformed
+    }//GEN-LAST:event_btnxoaActionPerformed
 
-    private void jButton25ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton25ActionPerformed
+    private void btndoimkActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btndoimkActionPerformed
         // TODO add your handling code here:
         new DoiMatKhau().setVisible(true);
 
-    }//GEN-LAST:event_jButton25ActionPerformed
+    }//GEN-LAST:event_btndoimkActionPerformed
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
         // TODO add your handling code here:
@@ -394,7 +410,7 @@ public class TrangNhanVien extends javax.swing.JFrame {
             pm.setString(1, txttim.getText());
             ResultSet rs = pm.executeQuery();
             if (rs.next()) {
-
+                model.setRowCount(1);
                 int s5 = rs.getInt(5);
                 String chucvu;
                 if (s5 == 1) {
@@ -402,33 +418,40 @@ public class TrangNhanVien extends javax.swing.JFrame {
                 } else {
                     chucvu = "Nhan Vien";
                 }
-                JOptionPane.showMessageDialog(this, "Tim thay manv " + tblnv.getValueAt(index, 1) + "\n" + "Ho ten " + tblnv.getValueAt(index, 2)
-                        + "\n" + "Ngay sinh " + tblnv.getValueAt(index, 3) + "\n" + "Dien thoai " + tblnv.getValueAt(index, 4)
-                        + "\n" + "Chuc vu " + chucvu
-                        + "\n" + "So CMT " + tblnv.getValueAt(index, 6));
+                tblnv.setValueAt(rs.getString(1), index, 1);
+                tblnv.setValueAt(rs.getString(2), index, 2);
+                tblnv.setValueAt(rs.getString(3), index, 3);
+                tblnv.setValueAt(rs.getString(4), index, 4);
+                tblnv.setValueAt(chucvu, index, 5);
+                tblnv.setValueAt(rs.getString(7), index, 6);
+
+//                JOptionPane.showMessageDialog(this, "Tim thay manv " + tblnv.getValueAt(index, 1) + "\n" + "Ho ten " + tblnv.getValueAt(index, 2)
+//                        + "\n" + "Ngay sinh " + tblnv.getValueAt(index, 3) + "\n" + "Dien thoai " + tblnv.getValueAt(index, 4)
+//                        + "\n" + "Chuc vu " + chucvu
+//                        + "\n" + "So CMT " + tblnv.getValueAt(index, 6));
             } else {
                 JOptionPane.showMessageDialog(this, "Không tìm thấy người học này!");
             }
-            model.setRowCount(1);
-            int s5 = rs.getInt(5);
-            String chucvu;
-            if (s5 == 1) {
-                chucvu = "Quan Ly";
-            } else {
-                chucvu = "Nhan Vien";
-            }
-            tblnv.setValueAt(rs.getString(1), index, 1);
-            tblnv.setValueAt(rs.getString(2), index, 2);
-            tblnv.setValueAt(rs.getString(3), index, 3);
-            tblnv.setValueAt(rs.getString(4), index, 4);
-            tblnv.setValueAt(chucvu, index, 5);
-            tblnv.setValueAt(rs.getString(7), index, 6);
+
             rs.close();
             pm.close();
 
         } catch (Exception e) {
         }
     }//GEN-LAST:event_jButton6ActionPerformed
+
+    private void tblnvMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tblnvMouseClicked
+        // TODO add your handling code here:
+        try {
+            index = tblnv.getSelectedRow();
+            if (index >= 0) {
+//                NhanVien s = lstnv.get(index);
+               
+                showdetail();
+            }
+        } catch (Exception e) {
+        }
+    }//GEN-LAST:event_tblnvMouseClicked
 
     /**
      * @param args the command line arguments
@@ -466,12 +489,12 @@ public class TrangNhanVien extends javax.swing.JFrame {
     }
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton btnadd;
+    private javax.swing.JButton btndoimk;
+    private javax.swing.JButton btnupdate;
+    private javax.swing.JButton btnxoa;
     private javax.swing.JButton jButton1;
     private javax.swing.JButton jButton2;
-    private javax.swing.JButton jButton22;
-    private javax.swing.JButton jButton23;
-    private javax.swing.JButton jButton24;
-    private javax.swing.JButton jButton25;
     private javax.swing.JButton jButton26;
     private javax.swing.JButton jButton27;
     private javax.swing.JButton jButton3;
@@ -480,11 +503,11 @@ public class TrangNhanVien extends javax.swing.JFrame {
     private javax.swing.JButton jButton6;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel3;
     private javax.swing.JPanel jPanel4;
     private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JTextField lblcurrrent;
     private javax.swing.JTable tblnv;
     private javax.swing.JTextField txttim;
     // End of variables declaration//GEN-END:variables
@@ -511,7 +534,33 @@ private void loadtable() {
             JOptionPane.showMessageDialog(this, "Loi load table");
         }
     }
-private void check(){
-    JOptionPane.showMessageDialog(this, "YH");
-}
+private void loaddata() {
+        try {
+            String sql = "select * from nhanvien";
+            Statement stm = cn.createStatement();
+            ResultSet rs = stm.executeQuery(sql);
+            int stt = 1;
+            while (rs.next()) {
+                int s5 = rs.getInt(5);
+                String chucvu;
+                if (s5 == 1) {
+                    chucvu = "Quan Ly";
+                } else {
+                    chucvu = "Nhan Vien";
+                }
+                    lstnv.add(new NhanVien(rs.getString(1), rs.getString(2),rs.getDate(3), rs.getString(4) ,
+                           rs.getBoolean(5), rs.getString(6),rs.getString(7)));
+                
+            }
+            rs.close();
+            stm.close();
+        } catch (Exception e) {
+            JOptionPane.showMessageDialog(this, "Loi load table");
+        }
+    }
+
+    private void showdetail() {
+     NhanVien s = lstnv.get(index);
+     lblcurrrent.setText("Tổng số bản ghi: " + (index + 1) + "/" + lstnv.size());
+    }
 }
